Add palette::from_csv
