+++ src/user/vga.rs
@@ -1,7 +1,6 @@
 use crate::{api, kernel, print, user};
-use crate::api::vga::Palette;
+use crate::api::vga::palette;
 use alloc::vec::Vec;
-use core::convert::TryInto;
 
 pub fn main(args: &[&str]) -> user::shell::ExitCode {
     if args.len() == 1 {
@@ -25,27 +24,13 @@ pub fn main(args: &[&str]) -> user::shell::ExitCode {
                 }
             } else if args.len() == 4 && args[2] == "palette" {
                 if let Some(file) = kernel::fs::File::open(args[3]) {
-                    let mut colors = Vec::with_capacity(16);
-                    for line in file.read_to_string().split("\n") {
-                        let line = line.split("#").next().unwrap();
-                        let color: Vec<u8> = line.split(",").filter_map(|value| {
-                            let radix = if value.contains("0x") { 16 } else { 10 };
-                            let value = value.trim().trim_start_matches("0x");
-                            u8::from_str_radix(value, radix).ok()
-                        }).collect();
-                        if color.len() == 4 {
-                            colors.push((color[0], color[1], color[2], color[3]));
-                        }
-                    }
-                    if let Ok(colors) = colors.try_into() {
-                        let palette = Palette { colors };
+                    if let Ok(palette) = palette::from_csv(&file.read_to_string()) {
                         kernel::vga::set_palette(palette);
                     } else {
                         print!("Could not parse palette file\n");
                         return user::shell::ExitCode::CommandError;
                     }
                 }
-
             } else {
                 print!("Invalid command\n");
                 return user::shell::ExitCode::CommandError;
